+++ packages/backend-common/src/tokens/types.ts
@@ -20,6 +20,20 @@
  * @public
  */
 export interface TokenManager {
-  getToken: () => Promise<{ token: string }>;
-  authenticate: (token: string) => Promise<void>;
+  /**
+   * Fetches a valid token.
+   *
+   * @remarks
+   *
+   * Tokens are valid for roughly one hour; the actual deadline is set in the
+   * payload `exp` claim. Never hold on to tokens for reuse; always ask for a
+   * new one for each outgoing request. This ensures that you always get a
+   * valid, fresh one.
+   */
+  getToken(): Promise<{ token: string }>;
+
+  /**
+   * Validates a given token.
+   */
+  authenticate(token: string): Promise<void>;
 }