+++ plugins/kubernetes-common/src/types.ts
@@ -31,6 +31,9 @@ import { Entity } from '@backstage/catalog-model';
 export interface KubernetesRequestBody {
   auth?: {
     google?: string;
+    oidc?: {
+      [key: string]: string;
+    };
   };
   entity: Entity;
 }