+++ plugins/techdocs-addons-test-utils/src/test-utils.tsx
@@ -46,6 +46,7 @@ const techdocsApi = {
 
 const techdocsStorageApi = {
   getApiOrigin: jest.fn(),
+  getBaseUrl: jest.fn(),
   getEntityDocs: jest.fn(),
   syncEntityDocs: jest.fn(),
 };
@@ -64,14 +65,18 @@ type TechDocsAddonTesterTestApiPair<TApi> = TApi extends infer TImpl
   : never;
 
 /** @ignore */
-type TechdocsAddonTesterApis<T> = TechDocsAddonTesterTestApiPair<T>[];
+type TechdocsAddonTesterApis<TApiPairs> = {
+  [TIndex in keyof TApiPairs]: TechDocsAddonTesterTestApiPair<
+    TApiPairs[TIndex]
+  >;
+};
 
 type TechDocsAddonTesterOptions = {
   dom: ReactElement;
   entity: Partial<TechDocsEntityMetadata>;
   metadata: Partial<TechDocsMetadata>;
   componentId: string;
-  apis: TechdocsAddonTesterApis<any>;
+  apis: TechdocsAddonTesterApis<any[]>;
   path: string;
 };
 
@@ -108,22 +113,40 @@ const defaultDom = (
 );
 
 /**
+ * Utility class for rendering TechDocs Addons end-to-end within the TechDocs
+ * reader page, with a set of givens (e.g. page DOM, metadata, etc).
+ *
+ * @example
+ * ```tsx
+ * const { getByText } = await TechDocsAddonTester.buildAddonsInTechDocs([<AnAddon />])
+ *   .withDom(<body>TEST_CONTENT</body>)
+ *   .renderWithEffects();
+ *
+ * expect(getByText('TEST_CONTENT')).toBeInTheDocument();
+ * ```
+ *
  * @public
  */
-
 export class TechDocsAddonTester {
   private options: TechDocsAddonTesterOptions = defaultOptions;
   private addons: ReactElement[];
 
+  /**
+   * Get a TechDocsAddonTester instance for a given set of Addons.
+   */
   static buildAddonsInTechDocs(addons: ReactElement[]) {
     return new TechDocsAddonTester(addons);
   }
 
-  private constructor(addons: ReactElement[]) {
+  // Protected in order to allow extension but not direct instantiation.
+  protected constructor(addons: ReactElement[]) {
     this.addons = addons;
   }
 
-  withApis<T>(apis: TechdocsAddonTesterApis<T>) {
+  /**
+   * Provide mock API implementations if your Addon expects any.
+   */
+  withApis<T extends any[]>(apis: TechdocsAddonTesterApis<T>) {
     const refs = apis.map(([ref]) => ref);
     this.options.apis = this.options.apis
       .filter(([ref]) => !refs.includes(ref))
@@ -131,28 +154,46 @@ export class TechDocsAddonTester {
     return this;
   }
 
+  /**
+   * Provide mock HTML if your Addon expects it in the shadow DOM.
+   */
   withDom(dom: ReactElement) {
     this.options.dom = dom;
     return this;
   }
 
+  /**
+   * Provide mock techdocs_metadata.json values if your Addon needs it.
+   */
   withMetadata(metadata: Partial<TechDocsMetadata>) {
     this.options.metadata = metadata;
     return this;
   }
 
+  /**
+   * Provide a mock entity if your Addon needs it. This also controls the base
+   * path at which the Addon is rendered.
+   */
   withEntity(entity: Partial<TechDocsEntityMetadata>) {
     this.options.entity = entity;
     return this;
   }
 
+  /**
+   * Provide the TechDocs page path at which the Addon is rendered (e.g. the
+   * part of the path after the entity namespace/kind/name).
+   */
   atPath(path: string) {
     this.options.path = path;
     return this;
   }
 
+  /**
+   * Return a fully configured and mocked TechDocs reader page within a test
+   * App instance, using the given Addon(s).
+   */
   build() {
-    const apis: TechdocsAddonTesterApis<any> = [
+    const apis: TechdocsAddonTesterApis<any[]> = [
       [techdocsApiRef, techdocsApi],
       [techdocsStorageApiRef, techdocsStorageApi],
       [searchApiRef, searchApi],
@@ -179,7 +220,9 @@ export class TechDocsAddonTester {
     techdocsStorageApi.getApiOrigin.mockResolvedValue(
       'https://backstage.example.com/api/techdocs',
     );
-
+    techdocsStorageApi.getBaseUrl.mockResolvedValue(
+      `https://backstage.example.com/api/techdocs/${entityName.namespace}/${entityName.kind}/${entityName.name}/${this.options.path}`,
+    );
     techdocsStorageApi.getEntityDocs.mockResolvedValue(
       renderToStaticMarkup(this.options.dom || defaultDom),
     );
@@ -215,11 +258,19 @@ export class TechDocsAddonTester {
     });
   }
 
-  // Components using useEffect to perform an asynchronous action (such as fetch) must be rendered within an async
-  // act call to properly get the final state, even with mocked responses. This utility method makes the signature a bit
-  // cleaner, since act doesn't return the result of the evaluated function.
-  // https://github.com/testing-library/react-testing-library/issues/281
-  // https://github.com/facebook/react/pull/14853
+  /**
+   * Render the Addon within a fully configured and mocked TechDocs reader.
+   *
+   * @remarks
+   * Components using useEffect to perform an asynchronous action (such as
+   * fetch) must be rendered within an async act call to properly get the final
+   * state, even with mocked responses. This utility method makes the signature
+   * a bit cleaner, since act doesn't return the result of the evaluated
+   * function.
+   *
+   * @see https://github.com/testing-library/react-testing-library/issues/281
+   * @see https://github.com/facebook/react/pull/14853
+   */
   async renderWithEffects(): Promise<
     typeof screen & { shadowRoot: ShadowRoot | null }
   > {