+++ packages/integration/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/integration
 
+## 1.2.0-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.1-next.0
+
 ## 1.2.0-next.0
 
 ### Minor Changes