+++ plugins/kafka-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-kafka-backend
 
+## 0.2.25-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.13.3-next.2
+  - @backstage/config@1.0.1-next.0
+  - @backstage/catalog-model@1.0.2-next.0
+
 ## 0.2.25-next.0
 
 ### Patch Changes